Moved datastore id generator to globals to prevent import loop
